Aligned Add control button to the right on the controls page.
Couldn’t find a way how to use different offset attribute on columns so went with adding an empty column that takes space on only the screen sizes where it’s needed.

Also changed the widths of the buttons on lower screen resolutions so they would look a bit nicer.